requirements.txt
flask
flask-smorest
flask_sqlalchemy